Improve the search field
Increase a bit the height and the font size in the search field to
make it more visible and accessible.

Also stylize it a bit with a drop shadow and a search icon (full
credit to octicons for the icon).

Finally, remove useless markup that were used back in the days to
support old browsers in order to vertically align the field.