Drastically speed up the GitHub related methods
We can drastically reduce the generation time by computing once data
that should not change (e.g. the repository URL).

We can also avoid going through any `if`, if the project is not in
a Git repository.

Finally, there's a tiny change in the last commit computation. The
generator used to pick the last commit that changed the current file
but we can safely take the repository's last commit to generate
permalinks.