Use turbolinks to make search panel persistent
By abandoning frames the side panel no longer persists after clicking on
links in the panel. With turbolinks we can make the panel persistent
across requests by setting data-turbolinks-permanent attribute on the
panel.

Generated pages can be in multiple sub directories so we link to the
assets with relative urls. Turbolinks does not consider relative urls
similar, resulting in page reloads and losing the context of the panel.
This has been fixed by making the assets urls absolute. This breaks
opening the docs locally. Maybe all generated docs should be in the same
folder?

This change also upgrade jQuery to 3.5.1.
Removes ie hack