Make panel responsive for mobile
On mobile we show the whole page including the search panel.
We can hide the panel and toggle it with css using a hidden checkbox.

This also changes the search input postion from fixed to sticky. This
will make sure the search input and close button are stick to the top.