Make sidepanel work with relative paths/URLs
With the migration to a frameless layout, the URLs to assets were made
absolute instead of relative because of Turbolinks asset tracking.
But absolute URLs fail when we have multiple versions of the docs in
subdirectories like: https://api.rubyonrails.org/v6.1/
See: https://github.com/rails/rails/issues/44042

Turbolinks tracking requires absolute URLs
-------------------------------------------

Turbolinks does not consider relative URLs to assets identical when
tracking assets for reload, even if the absolute URL would be the same.
For example when visiting /index.html, in the following css link:

    <link rel="stylesheet" href="css/panel.css" type="text/css" data-turbolinks-track="reload" />

... the panel.css would not be seen as identical as when visiting
/classes/Array.html with the following css link:

    <link rel="stylesheet" href="../css/panel.css" type="text/css" data-turbolinks-track="reload" />

This results in page reloads as Turbolinks wants to make sure it has all
the latests assets loaded. With a page reload we lose the context of the
panel when using search.

Instead of using absolute URLs another solution would be to remove the
`data-turbolinks-track` attribute.

Removing the data-turbolinks-track attribute
---------------------------------------------

By removing the `data-turbolinks-track` attribute we can use relative
URLs again. In this case Turbolinks just reloads the assets everytime
we navigate to another page. Reloading assets has a small performance
cost, but this seems barely noticable. It also requires a change to search results.

Fixing the search results URLs
------------------------------

The search results are persisted across navigations using the
`data-turbolinks-permanent` attribute. When we no longer track the assets,
the search tree gets rebuild for every navigation. But the URLs to the
search results are no longer available in the panel DOM as these are
stored in jQuery data attributes.
jQuery does not persist this data when navigating to another page:

    The data- attributes are pulled in the first time the data property
    is accessed and then are no longer accessed or mutated (all data
    values are then stored internally in jQuery).

    https://api.jquery.com/data/#data1

By storing the path in a HTML data-attribute it starts working again.

Updating the rel_prefix
-----------------------

When we use relative URLs the rel_prefix will change when we navigate to
a page that is on another level. For example, navigating from
`/classes/ActiveRecord.html` to `/classes/ActiveRecord/Attributes/ClassMethods.html`
changes the prefix from `../` to `../../../`.
If the prefix can change on every navigation, the prefix needs to be
loaded from the meta tag everytime, instead of on the first
initialization only.