Allow setting CSS based version badge
Currently the Rails API documentation inserts a HTML fragment containing
an image to add the Edge badge:
https://github.com/rails/rails-docs-server/blob/c8ca05fc9e434618a8e74b805c5b3fb94cf83caa/lib/generators/main.rb#L29-L45

Instead we can reuse the CSS-based Guides badge.
This also allows us to add the badge for stable versions, just like we
do for the guides.
The badge will only be shown when passing in the HORO_BADGE_VERSION env
variable.