Infer title and heading from comment
Some classes in Rails have a h1 heading defined in the RDoc comment.
For example:
  https://github.com/rails/rails/blob/4e1fe73028f4b01c8e5179989511b21925f70b20/activesupport/lib/active_support/current_attributes.rb#L8
  https://github.com/rails/rails/blob/4e1fe73028f4b01c8e5179989511b21925f70b20/actionpack/lib/action_controller/metal/strong_parameters.rb#L1159
This makes the heading more readable especially when there are a lot of
namespaces.

For SEO it helps if all classes have a h1 heading, but having to add
them by hand is cumbersome when we can generate them instead.

By looking at the comment of the class/module we can see if it has a
`h1` heading when it's starting with `= ` or `# `. If the heading is
missing we can generate a heading with the full name of the class
instead. This will prevent us from having to add the h1 for each class
in Rails.

This also adds the `@options.title` to the `title` tag of every class.