Replace highlight.js with Rouge gem
Using Rouge allows us to match the syntax highlighting style of the
Rails guides.  We can reuse the same stylesheet, and we can tweak the
language detection heuristics to match the majority of code snippets in
the Rails API docs.

Using Rouge also saves rendering time on the client side.