Replace rel_prefix with <base> tag + postprocessing
The Rails logo in the `data-turbo-permanent` part of the navigation
panel uses a path that is static after the first page load.  #228 made
the path relative, but it is still static.  This means that, when
navigating to a different page, the relative path is re-resolved using
that page's URL.

For example, when first visiting `foo/bar.html`, the image URL will be
`../i/logo.svg` which resolves to `/i/logo.svg`.  Then, after clicking
a link to `foo/bar/qux.html`, the image URL will still be `../i/logo.svg`
but will resolve to `foo/i/logo.svg`, causing the image to break.

More generally, any relative URLs in a `data-turbo-permanent` element
will break when navigating to a different directory level.  This hinders
us from adding static `<a>` elements to the navigation panel as well.
(The current navigation tree sidesteps this problem by using JavaScript
click handlers instead of actual links, but that has its own issues.)

As a solution, this commit replaces the use of `rel_prefix` with a
`<base>` tag and coordinated postprocessing of URLs.

The `<base>` tag for each page uses the (former) `rel_prefix` for that
page.  For example, the `<base>` tag for `foo/bar.html` would be
`<base href="../">`, and for `foo/bar/qux.html` would be
`<base href="../../">`.

Postprocessing then converts all URLs on each page so that they can be
resolved relative to the `<base>` tag.

URLs that previously used `rel_prefix` are now written as absolute
paths.  Postprocessing converts them to relative paths.  For example,
`#{rel_prefix}/i/logo.svg` is now written as just `/i/logo.svg`, which
postprocessing converts to `i/logo.svg`, which is resolved as
`../i/logo.svg` in `foo/bar.html` or `../../i/logo.svg` in
`foo/bar/qux.html`.

This fixes the disappearing logo, and allows us to put links and images
inside the `data-turbo-permanent` part of the navigation panel without
additional JavaScript.

This also addresses the issue that #176 describes, wherein Turbo will
reload assets when they have a different relative URL, even if they
would resolve to the same absolute URL.  Because asset paths are now the
same for all pages, Turbo will no longer reload them.