Enable horizontal scrollbar for method source code
When method source code is shown and contains long lines, it can
overflow and cause the entire page to be horizontally scrollable.

To prevent that, this commit sets `overflow-x: auto` for method source
code blocks (the same as for other code blocks), so that only the block
is horizontally scrollable.