Replace source code toggler JavaScript with <details>
This commit replaces the `javascript:toggleSource(...)` links used to
show and hide method source with an inherently toggleable `<details>`
element.  In addition to reducing JavaScript, this fixes a small
horizontal shift that occurred when changing "show" to "hide" (and vice
versa).

The `external-link.svg` file is from [Feather icons][feather] v4.29.0,
and is [licensed under the MIT license][license].

[feather]: https://feathericons.com/
[license]: https://github.com/feathericons/feather/blob/v4.29.0/LICENSE