Demo the issue solved by #240
This demonstrates the issue that #240 solves.

Netlify converts image URLs to absolute (CDN) URLs, so the disappearing
logo problem does not appear in the Netlify doc previews.

To properly demonstrate the problem, this commit adds a link to the
`data-turbo-permanent` part of the navigation panel.  The link uses
`rel_prefix` the same way the logo image does (but Netlify doesn't
rewrite the link URL).

If you click on the link when you first visit the doc preview, it will
work.  But if you visit another page first and then click the link, it
will 404.  Similarly, if you click the link twice, the second time will
404.

See #240 for an explanation.