Include site title in page titles
This adds the site title to each page's title.  When using the default
site title, this will incorporate `ENV["HORO_PROJECT_NAME"]` and
`ENV["HORO_BADGE_VERSION"]` into the page title.  For example,

  ```html
  <title>ActiveRecord::Base</title>
  ```

will become

  ```html
  <title>ActiveRecord::Base - Ruby on Rails v7.0.0 API documentation</title>
  ```

The intent is to give search engines more context and to differentiate
between versions of the documentation.