Prefer ENV["HORO_BADGE_VERSION"] version in banner
Prior to this commit, the banner at the top of each page would display
the value of `ENV["HORO_PROJECT_VERSION"]`.  For Rails, this would
either be a bare number, e.g. "7.0.0" without a "v" prefix, or a commit
reference, e.g. `main@1234567`.

In contrast, the value of `ENV["HORO_BADGE_VERSION"]` is intended to be
more user friendly.  For Rails, it would be e.g. "v7.0.0" or "edge".

This commit changes the banner to display `ENV["HORO_BADGE_VERSION"]`
if it is set, and otherwise fall back to `ENV["HORO_PROJECT_VERSION"]`.
However, since it may still be valuable for maintainers to know the
exact commit that the documentation was generated from, this commit also
adds a tooltip to show the abbreviated SHA1 of the commit when hovering
over the version.