Redesign banner and mobile menu bar
This commit redesigns the banner that was formerly displayed at the top
of each page.  The banner and the logo are now integrated, and they also
serve as the mobile menu bar.

The `menu.svg` and `close.svg` files are from [Feather icons][feather]
v4.29.0, and are [licensed under the MIT license][license].

[feather]: https://feathericons.com/
[license]: https://github.com/feathericons/feather/blob/v4.29.0/LICENSE