Remove banner from index.rhtml
The banner on the index page contains superfluous and even misleading
information.  The project name should be conveyed by the page title and
logo, and the project version should be conveyed by the page title and
version badge.  Furthermore, the "Last modified" time is the mtime of
the file on disk, unrelated to the time of the last commit for the file
or even for the branch as a whole.

This commit removes the banner from the index page for a cleaner
appearance.