Use pixels for `@media` queries
These `@media` queries are concerned with displaying a side panel that
is explicitly 300px wide.  Therefore, they should be expressed in terms
of pixels instead of variable-width units.  Also note that `em` in a
`@media` query is based on the default font size of the browser
configured by the user, _not_ the font size of `<body>`, which seems to
have been the intention with these (40 * 15px == 600px => equal space
for panel and content).