Unsticky "Appears in" element from bottom of page
Stickying the "Appears in" `<details>` at the bottom of the page causes
the element to jump up when it is expanded, which is unexpected and
unusual.  Since the element already hides information by default, there
is no need to force it to stay all the way at the bottom of the page.
This commit removes the sticky positioning and fixes some
inconsistencies in the grid definition.