Redesign "Appears in" file list
The "Appears in" file list is collapsed by default to avoid overwhelming
the user with very long file lists.  For example, the `Rails` module
appears in 125 files.  However the vast majority of modules appear in
only in their own file.  And, out of all 1370 modules, only ~50 appear
in 10 or more files:

  ```console
  $ grep -c '[.]rb</a>' doc/public/classes/**/*.html | head
  doc/rdoc/classes/AbstractController/ActionNotFound.html:1
  doc/rdoc/classes/AbstractController/Base.html:1
  doc/rdoc/classes/AbstractController/Caching/ClassMethods.html:1
  doc/rdoc/classes/AbstractController/Caching/ConfigMethods.html:1
  doc/rdoc/classes/AbstractController/Caching/Fragments/ClassMethods.html:1
  doc/rdoc/classes/AbstractController/Caching/Fragments.html:1
  doc/rdoc/classes/AbstractController/Caching.html:2
  doc/rdoc/classes/AbstractController/Callbacks/ClassMethods.html:1
  doc/rdoc/classes/AbstractController/Callbacks.html:1
  doc/rdoc/classes/AbstractController/Collector.html:1

  $ grep -c '[.]rb</a>' doc/public/classes/**/*.html | grep ':1$' | wc -l
  1181

  $ grep -c '[.]rb</a>' doc/public/classes/**/*.html | grep ':[1-5]$' | wc -l
  1293

  $ grep -c '[.]rb</a>' doc/public/classes/**/*.html | grep ':[1-9]$' | wc -l
  1321

  $ grep -c '[.]rb</a>' doc/public/classes/**/*.html | wc -l
  1370
  ```

This commit redesigns the "Appears in" file list to show a small number
of files by default.  If a module appears in 9 or fewer files, then all
the files will be shown by default.  Otherwise, the first 5 files will
shown, and a "More" button is shown to reveal the remaining files.