Add break opportunities in long names for mobile
When wrapping text, Safari and Chrome do not consider `/` nor `:` as
word break characters.  Firefox considers `/` as a word break character
but not `:`.  This can lead to horizontal overflow issues with long file
paths and module names on mobile.

This commit adds a `full_name` helper to insert word-break opportunities
into file paths and module names.  It also revises the `link_to` helper
to do the same when given an `RDoc::CodeObject`.  The helpers insert
`<wbr>` tags which tell the browser that the text can be wrapped at that
point.