Respect user font preferences
Technically savvy users are more likely to have a font preference
configured, and developers are more likely to have a preference for a
particular monospace font.

This commit removes the font family and base font size from the
stylesheet so that users can view the documentation with their preferred
fonts.

In particular, this stops forcing the use of the Courier font for
`<code>` elements nested in `<p>` when the Consolas and Menlo fonts are
not installed, which is commonly the case on Linux.  This also fixes
mismatching fonts for `<code>` elements nested in `<p>` versus `<code>`
elements nested in `<pre>`.