Fix ref links in main page
The main page reuses an already-processed file.  All of the ref links in
that file will be rendered relative to its path rather than relative to
the root.  For example, ref links in `railties/RDOC_MAIN.md` will be
relative to `files/railties/`, so a link to `ActiveRecord::Base` would
use `../../ActiveRecord/Base.html` instead of `ActiveRecord/Base.html`.

This commit changes `RDoc::Generator::SDoc#index` to dup the processed
file and override its `path` value such that links on the main page use
the root-relative path (but links on the file's page will still use the
`files/`-relative path).

This change also makes it unnecessary to escape occurrences of "Rails"
in the main page because, with the correct `href`, the postprocessor
will now recognize them as unintentional ref links.