Contain more-less <summary> inside <details>
Prior to this commit, the `<summary>` element of a more-less list was
positioned just below the bottom of the `<details>` element, which
caused it overlap with any subsequent padding or margin.

This commit extends the `<details>` element by adding bottom padding,
and aligns the bottom of the `<summary>` element with the bottom of the
`<details>` element so that the `<summary>` is always contained within
the block.