Monkey patch ToHtmlCrossref to modify ref links
Prior to this commit, postprocessing was used to unlink unintentional
ref links and style unstyled code ref links.  Postprocessing only
affects the final generator output, which is generally more robust.  But
in these cases, it would be more consistent for intermediate consumers
of `RDoc::CodeObject#description` to also see these changes.

This commit instead monkey patches `RDoc::Markup::ToHtmlCrossref` so
that it returns bare text for unintentional ref links and wraps
unstyled code ref links with `<code></code>`.  Monkey patching RDoc is
not ideal, but it allows all consumers of `RDoc::CodeObject#description`
to see these changes.