Fix link to crawler file
The `generate_file_links` method generates a `panel/links.html` file for
search engines to crawl.  When `panel/index.html` was converted to
`_panel.rhtml` in 4c5d4a0863ef1c44a41e828ce621c23a6a2e2712, the link to
`panel/links.html` was broken.  Though the value of this file may be
debatable, it definitely provides zero value if it is unreachable.  This
commit fixes the broken link.

This commit also renames the `generate_search_index` method to
`generate_file_links`, and renames `search_index.rhtml` to
`file_links.rhtml` so that it won't be confused with `search_index.js`,
which is entirely unrelated.