Use ESM version of Turbo
Turbo triggers a `turbo:load` event immediately after it is loaded.  If
a deferred script (such as a module) adds a listener for `turbo:load`,
and Turbo itself is not deferred, then the listener will only be invoked
on subsequent visits.

This commit replaces the UMD version of Turbo with the ESM version so
that its loading is deferred.