Prevent git commands if project is not a git repo
Prior to this commit, git commands (such as the command to get
`remote.origin.url`) would always be run if git was installed.  This
caused "fatal: not a git repository" to be printed to `stderr` when the
project was not a git repository.

This commit modifies `SDoc::Helpers::Git#git?` such that git commands
will only be run if git is installed _and_ the project is a git
repository.