Conserve <h1> headings in prose files
For prose files, such as `README` files, it is reasonable to assume that
the file content includes an `<h1>` heading, and that the file name is
mostly irrelevant.  For example, the file `railties/README.rdoc`
includes the heading "Railties -- Gluing the Engine to the Rails".

This commit removes the `<h1>` for a prose file's name, and preserves
the file's original `<h1>` instead of making it a subheading of an
`<hgroup>` as is done for module descriptions.