Refactor rendering
This reimplements the `SDoc::Templatable` module as an `SDoc::Renderer`
class.  An `SDoc::Renderer` is instantiated per file render, which
provides an isolated scope for each rendering.  Utilizing that
isolation, the context object (the `RDoc::CodeObject` instance) is now
stored in `@context` so that it doesn't have to be explicitly passed to
partials.

Also, as an optimization, `SDoc::Renderer` memoizes compiled ERB.