Bring attention to method name in method signature
This wraps the name portion of a method's signature in a `<b>` tag,
including names in `:call-seq:` method signatures.

In modern HTML, the `<b>` tag is "bring attention to element" rather
than "bold".  It is conventionally rendered with bold font weight, but
it has no semantic meaning (unlike, for example, the `<strong>` tag).

Co-authored-by: Shouichi Kamiya <shouichi.kamiya@gmail.com>