Add --core-ext option
The `--core-ext` option can be used to indicate files that define core
extensions.  It accepts a regular expression pattern.  If a module is
defined exclusively in files that match the pattern, the module will be
designated as a core extension and will be listed separately in the main
page nav.

The default value for `--core-ext` is `/core_ext/` which matches Rails'
core extension files in `activesupport/lib/active_support/core_ext/`.