Prevent RDoc from polluting in_files
There is currently a bug in RDoc v6.5.0 that causes the value of
`RDoc::ClassModule#in_files` to become polluted when referring to a
constant nested in a module where the module has been previously seen
but the constant has not.  For example, if RDoc sees a core extension
like:

  ```ruby
  class Float
    def as_json(*)
      # ...
    end
  end
  ```

and later sees a reference to `Float::INFINITY` in a file `foo.rb`, then
it will add `foo.rb` to `in_files` for `Float`.  That will cause
`foo.rb` to appear in `public/classes/Float.html` and `Float` to appear
in `public/files/foo.rb`.

This commit monkey patches RDoc to prevent such pollution.  The monkey
patch has been tested with the test suites of RDoc v6.5.0 and RDoc
`master` as of ruby/rdoc@9e75bf69b8b99aef5259e4a816977811ce6dc6c6, and
all tests pass.