Rework method spotlight effect
a67ece6220f8da3ed7f3319b89550234d1728382 changed the HTML around method
permalinks, which broke the spotlight effect when navigating to a
specific method.  However, prior to that commit (since upgrading to
Turbo v7), the spotlight effect only worked for cross-page navigation,
not intra-page navigation.  This commit fixes the spotlight effect for
both cross-page and intra-page navigation.

Additionally, this commit changes the visual effect from a momentary
flash of yellow to a persistent mark in the left margin.  This ensures
that effect is not missed (for example, when opening the page as a
background tab), and is less jarring when using dark mode.