Remove scroll-to-top button
Since d0e2ead20f0ca159503b2871667452abcedd4c2e, the table of contents
for a page is now in the sidebar / mobile menu instead of at the top of
the page, which eliminates the major use case for a scroll-to-top
button.

Furthermore, there are alternative ways to scroll to the top of the
page.  On desktop, the user can click the scoll bar or press the
<kbd>Home</kbd> key.  On mobile, the UX is less consistent, but still
possible.  For example, in iOS, the user can double-tap the status bar,
whereas Android has its own gestures.  While a scroll-to-top button
could provide a unified UX, due to limited screenspace, it is even more
likely to obscure content and be in the way.

Also, it is worth pointing out that one of uBlock Origin's bundled
filter lists ("EasyList - Annoyances") [blocks scroll-to-top buttons][].
This suggests that many people do not consider such buttons helpful.

[blocks scroll-to-top buttons]: https://github.com/easylist/easylist/blob/2d0b2d2295c9aa8b51deb21ff77da6824dcc24df/fanboy-addon/fanboy_annoyance_general_hide.txt#L498