Use explicit class instead of :target
Turbo Drive breaks the browser's back / forward buttons when navigating
to a history entry that was not added by Turbo Drive.  (The URL in the
address bar changes, but the page itself does not.)  Therefore, this
commit abandons working around Turbo Drive to fix `:target` behavior,
and adds an explicit `.target` class in response to navigation events
instead.