Remove extra whitespace in method signature params
RDoc includes extra whitespace in a method's parameter list when the
the method's `def` statement is spread across multiple lines.  For
example, a `def` like:

  ```ruby
  def define_attribute(
    name,
    cast_type,
    default: NO_DEFAULT_PROVIDED,
    user_provided_default: true
  )
    # ...
  end
  ```

Would produce a parameter list like:

  ```
  ( name, cast_type, default: NO_DEFAULT_PROVIDED, user_provided_default: true )
  ```

This commit monkey-patches `RDoc::AnyMethod#params` to remove the extra
whitespace within the parentheses:

  ```
  (name, cast_type, default: NO_DEFAULT_PROVIDED, user_provided_default: true)
  ```

This is implemented as a monkey patch so that it fixes `params` for
rendered method docs as well as search index entries.