Extract basic :call-seq: params for search results
Prior to this commit, search results displayed `(...)` as the parameter
list for any `:call-seq:` method.  The reason being the `:call-seq:`
value may contain method overloads (i.e. multiple parameter lists) or
even arbitrary text.  However, it is common for the `:call-seq:` value
to be basic method signature with a single parameter list.  Therefore,
this commit handles that case and otherwise falls back to `(...)`.