Support permalinking constants
This adds a permalink next to each constant's name, similar to methods.
This also adds a "Constants" navigation section in the sidebar, similar
to the "Methods" navigation section.

The permalinks are implemented by monkeypatching `RDoc::Constant` to
provide an `aref` method, just as `RDoc::ClassModule` and
`RDoc::MethodAttr` already do.