Fix compatibility with RDoc >= 6.10.0
ruby/rdoc@4a5206ae564fb6c3ba78e200b776001785c30839 added a kwarg to
`RDoc::Markup::ToHtmlCrossref#cross_reference` which must be forwarded.

This commit adds a double-splat (`**`) to the `cross_reference` monkey
patch in order to forward that kwarg.  It also adds a splat (`*`) in
case more positional args with default values are added in the future.